use cast

+++ test/deino_all_to_all.cpp
@@ -40,7 +40,7 @@ int mpi3::main(int /*argc*/, char** /*argv*/, mpi3::communicator world) try {
 
 	assert( sb.size() % static_cast<std::size_t>(world.size()) == 0);
 
-	world.all_to_all_n(sb.data(), sb.size()/world.size(), rb.data());
+	world.all_to_all_n(sb.data(), sb.size()/static_cast<std::size_t>(world.size()), rb.data());
 //    do_all_to_all_n(world, sb.data(), sb.size(), rb.data());
 //  do_all_to_all_n(world, sb.data(), sb.size(), rb.data());
 
@@ -48,7 +48,7 @@ int mpi3::main(int /*argc*/, char** /*argv*/, mpi3::communicator world) try {
     std::copy(sb.begin(), sb.end(), std::ostream_iterator<int>(wout<<"sb = ", ", ")); wout<<std::endl;
     std::copy(rb.begin(), rb.end(), std::ostream_iterator<int>(wout<<"rb = ", ", ")); wout<<std::endl;
 
-	world.all_to_all_inplace_n(sb.data(), sb.size()/world.size());
+	world.all_to_all_inplace_n(sb.data(), sb.size()/static_cast<std::size_t>(world.size()));
 //	do_all_to_all_n(world, sb.data(), sb.size(), sb.data());
 //	world.all_to_all_n(sb.data(), sb.size()); //  , sb.data());
 	std::copy(sb.begin(), sb.end(), std::ostream_iterator<int>(wout<<"sb (inplace) = ", ", ")); wout<<std::endl;
